requirements.txt
  appnope==0.1.4
  asttokens==2.4.1
  certifi==2024.7.4
  chardet==5.2.0
  charset-normalizer==3.3.2
  comm==0.2.2
  debugpy==1.8.5
  decorator==5.1.1
  exceptiongroup==1.2.2
  executing==2.0.1
  idna==3.7
  importlib_metadata==8.2.0
  ipykernel==6.29.5
  ipython==8.18.1
  jedi==0.19.1
  jupyter_client==8.6.2
  jupyter_core==5.7.2
  matplotlib-inline==0.1.7
  nest-asyncio==1.6.0
  packaging==24.1
  parso==0.8.4
  pexpect==4.9.0
  platformdirs==4.2.2
  prompt_toolkit==3.0.47
  psutil==6.0.0
  ptyprocess==0.7.0
  pure_eval==0.2.3
  Pygments==2.18.0
  python-dateutil==2.9.0.post0
python-dotenv==1.0.1
  pyzmq==26.1.0
requests==2.32.3
  six==1.16.0
  stack-data==0.6.3
  tornado==6.4.1
  traitlets==5.14.3
  typing_extensions==4.12.2
  urllib3==2.2.2
  wcwidth==0.2.13
  zipp==3.19.2
